refactor(server): create dedicated eval router
- Create new eval.ts file in src/server/routes/
- Move eval-related routes from server.ts to eval.ts
- Update server.ts to use the new evalRouter
- Improve code organization and maintainability
