+++ src/server/server.ts
@@ -10,27 +10,14 @@ import readline from 'node:readline';
 import opener from 'opener';
 import { Server as SocketIOServer } from 'socket.io';
 import invariant from 'tiny-invariant';
-import { v4 as uuidv4 } from 'uuid';
 import { fromError } from 'zod-validation-error';
 import { createPublicUrl } from '../commands/share';
 import { VERSION } from '../constants';
 import { getDbSignalPath } from '../database';
 import { getDirectory } from '../esm';
-import type {
-  EvaluateSummaryV2,
-  EvaluateTestSuiteWithEvaluateOptions,
-  GradingResult,
-  Job,
-  Prompt,
-  PromptWithMetadata,
-  ResultsFile,
-  TestCase,
-  TestSuite,
-} from '../index';
-import promptfoo from '../index';
+import type { Prompt, PromptWithMetadata, TestCase, TestSuite } from '../index';
 import logger from '../logger';
 import Eval from '../models/eval';
-import EvalResult from '../models/evalResult';
 import telemetry from '../telemetry';
 import { TelemetryEventSchema } from '../telemetry';
 import { synthesizeFromTestSuite } from '../testCases';
@@ -40,18 +27,13 @@ import {
   listPreviousResults,
   readResult,
   getTestCases,
-  updateResult,
   getLatestEval,
   migrateResultsFromFileSystemToDatabase,
   getStandaloneEvals,
-  deleteEval,
-  writeResultsToDatabase,
 } from '../util';
+import { evalRouter } from './routes/eval';
 import { providersRouter } from './routes/providers';
 import { redteamRouter } from './routes/redteam';
-
-// Running jobs
-const evalJobs = new Map<string, Job>();
 
 // Prompts cache
 let allPrompts: PromptWithMetadata[] | null = null;
@@ -91,126 +73,6 @@ export function createApp() {
         };
       }),
     });
-  });
-
-  app.post('/api/eval/job', (req, res) => {
-    const { evaluateOptions, ...testSuite } = req.body as EvaluateTestSuiteWithEvaluateOptions;
-    const id = uuidv4();
-    evalJobs.set(id, { status: 'in-progress', progress: 0, total: 0, result: null });
-
-    promptfoo
-      .evaluate(
-        Object.assign({}, testSuite, {
-          writeLatestResults: true,
-          sharing: testSuite.sharing ?? true,
-        }),
-        Object.assign({}, evaluateOptions, {
-          eventSource: 'web',
-          progressCallback: (progress: number, total: number) => {
-            const job = evalJobs.get(id);
-            invariant(job, 'Job not found');
-            job.progress = progress;
-            job.total = total;
-            console.log(`[${id}] ${progress}/${total}`);
-          },
-        }),
-      )
-      .then(async (result) => {
-        const job = evalJobs.get(id);
-        invariant(job, 'Job not found');
-        job.status = 'complete';
-        job.result = await result.toEvaluateSummary();
-        console.log(`[${id}] Complete`);
-      });
-
-    res.json({ id });
-  });
-
-  app.get('/api/eval/job/:id', (req, res) => {
-    const id = req.params.id;
-    const job = evalJobs.get(id);
-    if (!job) {
-      res.status(404).json({ error: 'Job not found' });
-      return;
-    }
-    if (job.status === 'complete') {
-      res.json({ status: 'complete', result: job.result });
-    } else {
-      res.json({ status: 'in-progress', progress: job.progress, total: job.total });
-    }
-  });
-
-  app.patch('/api/eval/:id', (req, res) => {
-    const id = req.params.id;
-    const { table, config } = req.body;
-
-    if (!id) {
-      res.status(400).json({ error: 'Missing id' });
-      return;
-    }
-
-    try {
-      updateResult(id, config, table);
-      res.json({ message: 'Eval updated successfully' });
-    } catch {
-      res.status(500).json({ error: 'Failed to update eval table' });
-    }
-  });
-
-  app.post('/api/eval/:evalId/results/:id/rating', async (req, res) => {
-    const { id } = req.params;
-    const gradingResult = req.body as GradingResult;
-    const result = await EvalResult.findById(id);
-    invariant(result, 'Result not found');
-    result.gradingResult = gradingResult;
-    result.success = gradingResult.pass;
-    result.score = gradingResult.score;
-
-    await result.save();
-    res.json(result);
-  });
-
-  app.post('/api/eval', async (req, res) => {
-    const body = req.body;
-    try {
-      if (body.data) {
-        logger.debug('[POST /api/eval] Saving eval results (v3) to database');
-        const { data: payload } = req.body as { data: ResultsFile };
-        const id = await writeResultsToDatabase(
-          payload.results as EvaluateSummaryV2,
-          payload.config,
-        );
-        res.json({ id });
-      } else {
-        const incEval = body as unknown as Eval;
-        logger.debug('[POST /api/eval] Saving eval results (v4) to database');
-        const eval_ = await Eval.create(incEval.config, incEval.prompts || [], {
-          author: incEval.author,
-          createdAt: new Date(incEval.createdAt),
-          results: incEval.results,
-        });
-        logger.debug(`[POST /api/eval] Eval created with ID: ${eval_.id}`);
-
-        logger.debug(
-          `[POST /api/eval] Saved ${incEval.results.length} results to eval ${eval_.id}`,
-        );
-
-        res.json({ id: eval_.id });
-      }
-    } catch (error) {
-      console.error('Failed to write eval to database', error, body);
-      res.status(500).json({ error: 'Failed to write eval to database' });
-    }
-  });
-
-  app.delete('/api/eval/:id', async (req, res) => {
-    const { id } = req.params;
-    try {
-      await deleteEval(id);
-      res.json({ message: 'Eval deleted successfully' });
-    } catch {
-      res.status(500).json({ error: 'Failed to delete eval' });
-    }
   });
 
   app.get('/api/results/:id', async (req, res) => {
@@ -280,6 +142,7 @@ export function createApp() {
 
   app.use('/api/providers', providersRouter);
   app.use('/api/redteam', redteamRouter);
+  app.use('/api/eval', evalRouter);
 
   app.post('/api/telemetry', async (req, res) => {
     try {
